+++ src/extensions/internal.extension.ts
@@ -21,6 +21,7 @@ import {
   START,
   TBlackHole,
   TContext,
+  TLoadableChildLifecycle,
   YEAR,
 } from "..";
 
@@ -240,6 +241,7 @@ export class InternalDefinition {
      * Roughly speaking, what's the application doing? Mostly useful for debugging
      */
     phase: Phase;
+    lifecycleHooks: Map<string, TLoadableChildLifecycle>;
   };
   public utils = new InternalUtils();
